[이안] 스프링 카페 4주차 구현

+++ src/main/java/kr/codesqaud/cafe/board/service/BoardService.java
@@ -3,6 +3,8 @@ package kr.codesqaud.cafe.board.service;
 import kr.codesqaud.cafe.board.dto.PostResponse;
 import kr.codesqaud.cafe.board.dto.PostWriteForm;
 import kr.codesqaud.cafe.board.repository.BoardJdbcRepository;
+import kr.codesqaud.cafe.board.repository.CommentJdbcRepository;
+import kr.codesqaud.cafe.exception.ForbiddenException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,28 +13,37 @@ import java.util.stream.Collectors;
 @Service
 public class BoardService {
     private final BoardJdbcRepository boardJdbcRepository;
+    private final CommentJdbcRepository commentJdbcRepository;
 
-    public BoardService(BoardJdbcRepository boardJdbcRepository) {
+    public BoardService(BoardJdbcRepository boardJdbcRepository, CommentJdbcRepository commentJdbcRepository) {
         this.boardJdbcRepository = boardJdbcRepository;
+        this.commentJdbcRepository = commentJdbcRepository;
     }
 
     public void write(PostWriteForm postWriteForm) {
-        boardJdbcRepository.save(postWriteForm.toBoardPost());
+        boardJdbcRepository.save(postWriteForm.toEntity());
     }
 
     public void update(PostResponse postResponse) {
-        boardJdbcRepository.update(postResponse.toBoardPost());
+        boardJdbcRepository.update(postResponse.toEntity());
     }
 
     public void delete(Long postId) {
-        boardJdbcRepository.delete(postId);
+        if (commentJdbcRepository.getCommentCountByOtherWriter(postId) == 0) {
+            commentJdbcRepository.deleteAllByPostId(postId);
+            boardJdbcRepository.delete(postId);
+        } else {
+            throw new ForbiddenException("작성자와 다른 댓글 작성자가 존재하여 삭제할 수 없습니다.");
+        }
     }
 
     public PostResponse getPost(Long postId) {
-        return PostResponse.fromBoardPost(boardJdbcRepository.findByPostId(postId));
+        return PostResponse.from(boardJdbcRepository.findByPostId(postId));
     }
 
     public List<PostResponse> getPostList() {
-        return boardJdbcRepository.findAll().stream().map(PostResponse::fromBoardPost).collect(Collectors.toList());
+        return boardJdbcRepository.findAll().stream()
+                .map(PostResponse::from)
+                .collect(Collectors.toList());
     }
 }
